renamed customer.js
